requirements.txt
qrcode
Pillow
customtkinter